pyproject.toml

[project]
name = "mistral-ocr"
version = "0.1.0"
description = "A CLI and Streamlit tool to convert PDFs to Markdown using Mistral OCR."
readme = "README.md"
authors = [
    { name = "Samuel Mallet", email = "69467005+SuperMuel@users.noreply.github.com" },
]
requires-python = ">=3.12"
# Combine dependencies from both CLI and Streamlit parts
dependencies = [
    "mistralai>=1.5.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.1",
    "tqdm>=4.67.1",
    "typer[rich]>=0.15.2",  # Add [rich] extra for console output formatting
]

[project.scripts]
# This points to the main function in src/mistral_ocr/__init__.py
mistral-ocr = "mistral_ocr:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.9.10"]

# Optional: Ruff configuration if using it
[tool.ruff]
exclude = ["*.ipynb"]
